i386: virt: Enable Linux kernel loading
We factorize the PC Linux kernel loading routine into an i386
kernel-loader API, and call it from both PC and virt.
This allows us to support the kernel (--kernel) loading option.
