+++ hw/i386/Makefile.objs
@@ -1,5 +1,5 @@
 obj-$(CONFIG_KVM) += kvm/
-obj-y += cpu.o memory.o multiboot.o fw.o
+obj-y += cpu.o memory.o multiboot.o fw.o kernel-loader.o
 obj-y += pc.o
 obj-$(CONFIG_I440FX) += pc_piix.o
 obj-$(CONFIG_Q35) += pc_q35.o
