+++ hw/i386/virt/virt.c
@@ -31,6 +31,7 @@
 #include "hw/i386/acpi.h"
 #include "hw/i386/cpu-internal.h"
 #include "hw/i386/fw.h"
+#include "hw/i386/kernel-loader.h"
 #include "hw/i386/topology.h"
 #include "hw/i386/amd_iommu.h"
 #include "hw/i386/intel_iommu.h"
@@ -104,6 +105,7 @@ static void virt_machine_state_init(MachineState *machine)
     //MemoryRegion *ram;
     MachineClass *mc = MACHINE_GET_CLASS(machine);
     VirtMachineState *vms = VIRT_MACHINE(machine);
+    bool linux_boot = (machine->kernel_filename != NULL);
 
     /* NUMA stuff */
     vms->numa_nodes = nb_numa_nodes;
@@ -121,13 +123,39 @@ static void virt_machine_state_init(MachineState *machine)
 
     vms->apic_id_limit = cpus_init(machine, false);
 
+    kvmclock_create();
+
     fw_cfg = fw_cfg_init(machine, smp_cpus, mc->possible_cpu_arch_ids(machine), vms->apic_id_limit);
     rom_set_fw(fw_cfg);
     vms->fw_cfg = fw_cfg;
-
-    kvmclock_create();
-
     acpi_conf_virt_init(machine);
+
+    if (linux_boot) {
+        load_linux(MACHINE(vms), vms->acpi_configuration, fw_cfg);
+    }
+}
+
+static void virt_machine_instance_init(Object *obj)
+{
+}
+
+static void virt_machine_reset(void)
+{
+    CPUState *cs;
+    X86CPU *cpu;
+
+    qemu_devices_reset();
+
+    CPU_FOREACH(cs) {
+        cpu = X86_CPU(cs);
+
+        /* Reset APIC after devices have been reset to cancel
+         * any changes that qemu_devices_reset() might have done.
+         */
+        if (cpu->apic_state) {
+            device_reset(cpu->apic_state);
+        }
+    }
 }
 
 static void x86_nmi(NMIState *n, int cpu_index, Error **errp)
@@ -158,6 +186,7 @@ static const TypeInfo virt_machine_info = {
     .parent        = TYPE_MACHINE,
     .abstract      = true,
     .instance_size = sizeof(VirtMachineState),
+    .instance_init = virt_machine_instance_init,
     .class_size    = sizeof(VirtMachineClass),
     .class_init    = virt_class_init,
     .interfaces = (InterfaceInfo[]) {
@@ -193,6 +222,7 @@ static void virt_machine_class_init(MachineClass *mc)
     mc->cpu_index_to_instance_props = cpu_index_to_props;
     mc->get_default_cpu_node_id = cpu_get_default_cpu_node_id;
     mc->possible_cpu_arch_ids = cpu_possible_cpu_arch_ids;;
+    mc->reset = virt_machine_reset;
 }
 
 static void virt_2_12_machine_class_init(MachineClass *mc)
